+++ src/renderer/src/lib/settings/appSettingsConfig.ts
@@ -7,7 +7,8 @@ export const appSettingsConfig: AppSettingConfig[] = [
 		tooltip: 'setting.appLocale_tooltip',
 		options: [
 			{ id: 'en-US', literalName: 'English' },
-			{ id: 'ENGLISH', literalName: 'ENGLISH' },
+			{ id: 'ja-JP', literalName: '日本語' },
+			{ id: 'zh-TW', literalName: '中文（正體字）' },
 		],
 	},
 	{